+++ packages/hono-takibi/src/helper/allof.ts
@@ -1,7 +1,7 @@
 import { zodSchemaFromSubSchema } from '../generator/zod/helper/zod-schema-from-sub-schema.js'
 import type { Schema } from '../openapi/types.js'
 import { intersection } from '../utils/index.js'
-import { isNullableSchema } from '../validator/is-nullable-schema.js'
+import { isNullableSchema } from '../validator/index.js'
 
 /**
  * Converts an OpenAPI `allOf` schema into a Zod intersection schema.
