--- packages/hono-takibi/src/validator/is-ref-object.ts
@@ -1,19 +1,0 @@
-/**
- * Checks if a value is a non-null object (e.g., a potential `$ref` object).
- *
- * @param value - The value to check.
- * @returns `true` if the value is a non-null object.
- *
- * @example
- * ```ts
- * isRefObject({ $ref: '#/components/schemas/User' }) // true
- * isRefObject(null)                                  // false
- * isRefObject('text')                                // false
- * ```
- */
-export function isRefObject(value: unknown): value is {
-  $ref?: string
-  [key: string]: unknown
-} {
-  return typeof value === 'object' && value !== null
-}
