--- packages/hono-takibi/src/validator/is-unique-content-schema.test.ts
@@ -1,23 +1,0 @@
-import { describe, expect, it } from 'vitest'
-import { isUniqueContentSchema } from '.'
-
-// Test run
-// pnpm vitest run ./src/validator/is-unique-content-schema.test.ts
-
-describe('isUniqueContentSchema Test', () => {
-  it.concurrent('isUniqueContentSchema -> true', () => {
-    const result = isUniqueContentSchema(['application/json'], {
-      'application/json': { schema: { $ref: '#/components/schemas/Test' } },
-    })
-    const expected = true
-    expect(result).toBe(expected)
-  })
-  it.concurrent('isUniqueContentSchema -> false', () => {
-    const result = isUniqueContentSchema(['application/json', 'application/xml'], {
-      'application/json': { schema: { $ref: '#/components/schemas/Test' } },
-      'application/xml': { schema: { $ref: '#/components/schemas/Example' } },
-    })
-    const expected = false
-    expect(result).toBe(expected)
-  })
-})
