--- packages/hono-takibi/src/validator/is-unique-content-schema.ts
@@ -1,21 +1,0 @@
-import type { Content } from '../openapi/index.js'
-
-/**
- * Checks if all given content types share the same schema definition.
- *
- * @param contentTypes - Array of content type keys (e.g., ['application/json', 'application/xml']).
- * @param content - OpenAPI content object mapping content types to media objects.
- * @returns `true` if all specified content types refer to the same schema; otherwise `false`.
- *
- * @example
- * ```ts
- * isUniqueContentSchema(['application/json', 'application/xml'], {
- *   'application/json': { schema: { type: 'string' } },
- *   'application/xml': { schema: { type: 'string' } },
- * }) // true
- * ```
- */
-export function isUniqueContentSchema(contentTypes: string[], content: Content): boolean {
-  const schema = new Set(contentTypes.map((type) => JSON.stringify(content?.[type].schema)))
-  return schema.size === 1
-}
